Filter a matrix, returning either all continuous or all discrete characters

Filter a matrix, returning either all continuous or all discrete characters

filterMatrix(mat, charType, returnType = "discrete")

Arguments

mat
A matrix of data

charType
A vector of character types (perhaps from detectAllCharacters)

returnType
Either discrete or continuous

Value

Matrix with only discrete or continuous characters

Examples

data(anolis)
aType<-detectAllCharacters(anolis$dat)

#> Warning: Guessing that this is a discrete character based on repeated values

#> Warning: Guessing that this is a discrete character based on repeated values

filterMatrix(anolis$dat, aType, "discrete")

#>     ecomorph      island
#> 1         TG        Cuba
#> 2         TW        Cuba
#> 3         GB        Cuba
#> 4         TC  Hispaniola
#> 5         TC        Cuba
#> 6         TG        Cuba
#> 7         TC        Cuba
#> 8         GB  Hispaniola
#> 9         GB        Cuba
#> 10        TW        Cuba
#> 11         U        Cuba
#> 12         U        Cuba
#> 13        TG  Hispaniola
#> 14        GB  Hispaniola
#> 15        CG  Hispaniola
#> 16        CG        Cuba
#> 17        CG  Hispaniola
#> 18         U        Cuba
#> 19         U  Hispaniola
#> 20         U        Cuba
#> 21        TG        Cuba
#> 22        TG  Hispaniola
#> 23         T  Hispaniola
#> 24         T  Hispaniola
#> 25         U        Cuba
#> 26         U        Cuba
#> 27        TC  Hispaniola
#> 28         U  Hispaniola
#> 29        GB        Cuba
#> 30        TC  Hispaniola
#> 31        TG        Cuba
#> 32        TG Puerto Rico
#> 33        TG Puerto Rico
#> 34        GB        Cuba
#> 35        CG Puerto Rico
#> 36        GB        Cuba
#> 37        TG  Hispaniola
#> 38        TW  Hispaniola
#> 39         T  Hispaniola
#> 40        GB  Hispaniola
#> 41        CG        Cuba
#> 42         U  Hispaniola
#> 43         U  Hispaniola
#> 44        TC Puerto Rico
#> 45         U  Hispaniola
#> 46        CG Puerto Rico
#> 47        TC Puerto Rico
#> 48        TG        Cuba
#> 49         U        Cuba
#> 50        TW        Cuba
#> 51        TG Puerto Rico
#> 52        TG  Hispaniola
#> 53        GB  Hispaniola
#> 54        TG        Cuba
#> 55        TG        Cuba
#> 56        GB        Cuba
#> 57        TW  Hispaniola
#> 58        TC        Cuba
#> 59        TG        Cuba
#> 60        GB Puerto Rico
#> 61        TG     Jamaica
#> 62        TG  Hispaniola
#> 63         T        Cuba
#> 64         U        Cuba
#> 65        CG        Cuba
#> 66        GB        Cuba
#> 67        TG  Hispaniola
#> 68         T  Hispaniola
#> 69        TG        Cuba
#> 70         U  Hispaniola
#> 71        CG        Cuba
#> 72        TW Puerto Rico
#> 73        GB  Hispaniola
#> 74        TC     Jamaica
#> 75        GB        Cuba
#> 76        TC        Cuba
#> 77        TW        Cuba
#> 78        TW  Hispaniola
#> 79        GB Puerto Rico
#> 80        TC        Cuba
#> 81         U        Cuba
#> 82        GB Puerto Rico
#> 83         U        Cuba
#> 84        TG        Cuba
#> 85         U     Jamaica
#> 86        CG  Hispaniola
#> 87        TG        Cuba
#> 88        TG        Cuba
#> 89        GB  Hispaniola
#> 90        TW  Hispaniola
#> 91        TG  Hispaniola
#> 92        TC  Hispaniola
#> 93        CG        Cuba
#> 94        TG  Hispaniola
#> 95        TC Puerto Rico
#> 96        TW     Jamaica
#> 97        GB        Cuba
#> 98         U        Cuba
#> 99         T  Hispaniola
#> 100       TG  Hispaniola