Returns the character matrix or phylogeny from a treedata.table object

Returns the character matrix or phylogeny from a treedata.table object

pull.treedata.table(tdObject, type = "dat")

Arguments

tdObject
A treedata.table object

Value

A data.table or phylo object from the original treedata.table object

Examples

data(anolis)
td <- as.treedata.table(anolis$phy, anolis$dat)
pull.treedata.table(td, type = "phy")

#>            ahli         allogus     rubribarbus           imias          sagrei 
#>            ahli         allogus     rubribarbus           imias          sagrei 
#>         bremeri quadriocellifer      ophiolepis         mestrei           jubar 
#>         bremeri quadriocellifer      ophiolepis         mestrei           jubar 
#>      homolechis        confusus           guafe         garmani        opalinus 
#>      homolechis        confusus           guafe         garmani        opalinus 
#>         grahami     valencienni      lineatopus      reconditus       evermanni 
#>         grahami     valencienni      lineatopus      reconditus       evermanni 
#>       stratulus           krugi      pulchellus       gundlachi       poncensis 
#>       stratulus           krugi      pulchellus       gundlachi       poncensis 
#>           cooki    cristatellus    brevirostris        caudalis          marron 
#>           cooki    cristatellus    brevirostris        caudalis          marron 
#>        websteri       distichus        barbouri         alumina    semilineatus 
#>        websteri       distichus        barbouri         alumina    semilineatus 
#>         olssoni      etheridgei         fowleri       insolitus       whitemani 
#>         olssoni      etheridgei         fowleri       insolitus       whitemani 
#>       haetianus        breslini         armouri         cybotes         shrevei 
#>       haetianus        breslini         armouri         cybotes         shrevei 
#>   longitibialis         strahmi        marcanoi        baleatus       barahonae 
#>   longitibialis         strahmi        marcanoi        baleatus       barahonae 
#>        ricordii   eugenegrahami     christophei         cuvieri        barbatus 
#>        ricordii   eugenegrahami     christophei         cuvieri        barbatus 
#>          porcus  chamaeleonides       guamuhaya   altitudinalis        oporinus 
#>          porcus  chamaeleonides       guamuhaya   altitudinalis        oporinus 
#>        isolepis        allisoni        porcatus     argillaceus       centralis 
#>        isolepis        allisoni        porcatus     argillaceus       centralis 
#>         pumilis        loysiana         guazuma        placidus        sheplani 
#>         pumilis        loysiana         guazuma        placidus        sheplani 
#>         alayoni     angusticeps        paternus       alutaceus    inexpectatus 
#>         alayoni     angusticeps        paternus       alutaceus    inexpectatus 
#>       clivicola    cupeyalensis    cyanopleurus         alfaroi      macilentus 
#>       clivicola    cupeyalensis    cyanopleurus         alfaroi      macilentus 
#>       vanidicus     argenteolus          lucius        bartschi    vermiculatus 
#>       vanidicus     argenteolus          lucius        bartschi    vermiculatus 
#>        baracoae          noblei      smallwoodi    luteogularis       equestris 
#>        baracoae          noblei      smallwoodi    luteogularis       equestris 
#>       monticola   bahorucoensis dolichocephalus      hendersoni     darlingtoni 
#>       monticola   bahorucoensis dolichocephalus      hendersoni     darlingtoni 
#>        aliniger      singularis    chlorocyanus     coelestinus        occultus 
#>        aliniger      singularis    chlorocyanus     coelestinus        occultus 
#> 100 Levels: ahli alayoni alfaroi aliniger allisoni allogus ... whitemani

pull.treedata.table(td, type = "dat")

#> Error in tdObject[[1]]$tip.label: $ operator is invalid for atomic vectors