Returns the character matrix or phylogeny from a treedata.table object

Returns the character matrix or phylogeny from a treedata.table object

pull.treedata.table(tdObject, type = "dat")

Arguments

tdObject
A treedata.table object

type
Whether the output of the function is a tree ('type="phylo"') or a data.table ('type="dat"')

Value

A data.table or phylo object from the original treedata.table object

Examples

data(anolis)
td <- as.treedata.table(anolis$phy, anolis$dat)

#> Phylo object detected 

#> No tips were dropped from the original tree/dataset

pull.treedata.table(td, type = "phy")

#> 
#> Phylogenetic tree with 100 tips and 99 internal nodes.
#> 
#> Tip labels:
#> 	ahli, allogus, rubribarbus, imias, sagrei, bremeri, ...
#> 
#> Rooted; includes branch lengths.

pull.treedata.table(td, type = "dat")

#>            tip.label       tip.label      SVL   PCI_limbs   PCII_head
#>   1:            ahli            ahli 4.039125 -3.24828599  0.37225191
#>   2:         allogus         allogus 4.040138 -2.84557021  0.60011341
#>   3:     rubribarbus     rubribarbus 4.078469 -2.23834859  1.11997785
#>   4:           imias           imias 4.099687 -3.04891725  2.33203488
#>   5:          sagrei          sagrei 4.067162 -1.74105547  2.02282431
#>   6:         bremeri         bremeri 4.113371 -1.81361138  2.60675012
#>   7: quadriocellifer quadriocellifer 3.901619 -2.26789400  0.99092075
#>   8:      ophiolepis      ophiolepis 3.637962  0.79151174  1.45857603
#>   9:         mestrei         mestrei 3.987147 -2.60261906  1.27568612
#>  10:           jubar           jubar 3.952605 -2.22414593  0.94489985
#>  11:      homolechis      homolechis 4.032806 -2.74499346  0.87926009
#>  12:        confusus        confusus 3.938442 -2.49591732  0.16823269
#>  13:           guafe           guafe 3.877457 -2.61408710  0.65608676
#>  14:         garmani         garmani 4.769473 -0.77352640  0.93712494
#>  15:        opalinus        opalinus 3.838376 -1.77943710 -0.32453812
#>  16:         grahami         grahami 4.154274 -2.30565354 -1.91393689
#>  17:     valencienni     valencienni 4.321524  2.94241389 -0.88460072
#>  18:      lineatopus      lineatopus 4.128612 -2.43081291 -3.12552390
#>  19:      reconditus      reconditus 4.482607 -2.72704156 -0.21040657
#>  20:       evermanni       evermanni 4.165605 -2.52899245  0.12548098
#>  21:       stratulus       stratulus 3.869881 -1.62264165 -0.52957490
#>  22:           krugi           krugi 3.886500 -1.68879533 -0.83128181
#>  23:      pulchellus      pulchellus 3.799022  0.16238151 -2.33846103
#>  24:       gundlachi       gundlachi 4.188105 -2.64936274 -0.56251604
#>  25:       poncensis       poncensis 3.820378  0.52782664  1.24062035
#>  26:           cooki           cooki 4.091535 -2.22079009  0.05979850
#>  27:    cristatellus    cristatellus 4.189820 -3.33026331 -0.62225189
#>  28:    brevirostris    brevirostris 3.874155 -3.28900081  1.38425488
#>  29:        caudalis        caudalis 3.911743 -1.78270839  1.90870776
#>  30:          marron          marron 3.831810 -2.84709341 -0.12032028
#>  31:        websteri        websteri 3.916546 -2.50870054  0.97482786
#>  32:       distichus       distichus 3.928796 -3.84343343  0.83890749
#>  33:        barbouri        barbouri 3.663932 -0.87536237  1.34009434
#>  34:         alumina         alumina 3.588941  0.72166993  1.52514444
#>  35:    semilineatus    semilineatus 3.696631  0.18403236 -0.13872022
#>  36:         olssoni         olssoni 3.793899  0.81682668  3.87442813
#>  37:      etheridgei      etheridgei 3.657991 -3.81065955 -0.75340900
#>  38:         fowleri         fowleri 4.288780 -1.03667003  1.50383762
#>  39:       insolitus       insolitus 3.800471  5.03062671  0.24137669
#>  40:       whitemani       whitemani 4.097479 -2.54584901 -3.25974024
#>  41:       haetianus       haetianus 4.316542 -3.40402925 -3.78444664
#>  42:        breslini        breslini 4.051111 -2.97497529 -3.34153375
#>  43:         armouri         armouri 4.121684 -3.16993598 -4.08620441
#>  44:         cybotes         cybotes 4.210982 -3.11637922 -2.99620854
#>  45:         shrevei         shrevei 3.983003 -2.25952607 -3.55788235
#>  46:   longitibialis   longitibialis 4.242103 -3.52243048 -1.61670439
#>  47:         strahmi         strahmi 4.274271 -3.87340303 -1.01635836
#>  48:        marcanoi        marcanoi 4.079485 -2.84448243 -2.78644148
#>  49:        baleatus        baleatus 5.053056  0.73766495  1.79646028
#>  50:       barahonae       barahonae 5.076958  0.90731040  2.33222342
#>  51:        ricordii        ricordii 5.013963  0.58940470  1.34670464
#>  52:   eugenegrahami   eugenegrahami 4.128504 -4.21209140  4.83351955
#>  53:     christophei     christophei 3.884652 -2.70596332  1.70527559
#>  54:         cuvieri         cuvieri 4.875012 -1.18955124 -0.71983364
#>  55:        barbatus        barbatus 5.003946  2.25020229 -3.30818172
#>  56:          porcus          porcus 5.038034  3.50614537 -2.98667964
#>  57:  chamaeleonides  chamaeleonides 5.042349  2.95597428 -3.69231889
#>  58:       guamuhaya       guamuhaya 5.036953  3.10947396 -4.07973423
#>  59:   altitudinalis   altitudinalis 3.842994  2.86751469 -6.12805047
#>  60:        oporinus        oporinus 3.845670  3.05888289 -2.91565342
#>  61:        isolepis        isolepis 3.657088  3.05697642 -4.34018373
#>  62:        allisoni        allisoni 4.375390  2.03597967 -3.74252815
#>  63:        porcatus        porcatus 4.258991  1.93193251 -3.82953305
#>  64:     argillaceus     argillaceus 3.757869 -0.11307851 -1.48944374
#>  65:       centralis       centralis 3.697941  0.73073876 -0.41683597
#>  66:         pumilis         pumilis 3.466860  0.55404946 -0.42584152
#>  67:        loysiana        loysiana 3.701240  0.31536951 -1.08038145
#>  68:         guazuma         guazuma 3.763884  8.16566506 -0.60605865
#>  69:        placidus        placidus 3.773967  7.38332866  0.53958973
#>  70:        sheplani        sheplani 3.682924 10.27389091  2.27784409
#>  71:         alayoni         alayoni 3.815705  3.40888624 -1.78335850
#>  72:     angusticeps     angusticeps 3.788595  4.58711071 -1.97673831
#>  73:        paternus        paternus 3.802961  2.90637723 -0.95846093
#>  74:       alutaceus       alutaceus 3.554891  1.17338999 -0.61493813
#>  75:    inexpectatus    inexpectatus 3.537439  2.50607244 -0.52340367
#>  76:       clivicola       clivicola 3.758726 -1.05114421 -0.58903393
#>  77:    cupeyalensis    cupeyalensis 3.462014  2.72419698  0.48960424
#>  78:    cyanopleurus    cyanopleurus 3.630161  0.43328510  0.98930100
#>  79:         alfaroi         alfaroi 3.526655  2.58175921  0.77876335
#>  80:      macilentus      macilentus 3.715765  3.15177391  3.26626613
#>  81:       vanidicus       vanidicus 3.626206  4.04436760  1.08716320
#>  82:     argenteolus     argenteolus 3.971307 -2.92797328  2.61964493
#>  83:          lucius          lucius 4.198915 -4.38139803  1.25981123
#>  84:        bartschi        bartschi 4.280547 -3.25502168  0.80565652
#>  85:    vermiculatus    vermiculatus 4.802849  0.22783467  1.41411087
#>  86:        baracoae        baracoae 5.042780  0.76817392  0.06987127
#>  87:          noblei          noblei 5.083473 -0.09337994 -0.96088851
#>  88:      smallwoodi      smallwoodi 5.035096 -0.13770527 -1.15140405
#>  89:    luteogularis    luteogularis 5.101085  0.39614254  0.73584435
#>  90:       equestris       equestris 5.113994  1.05461517  0.70720387
#>  91:       monticola       monticola 3.770613 -3.25118016  0.15599344
#>  92:   bahorucoensis   bahorucoensis 3.827445 -0.04747332 -2.55694724
#>  93: dolichocephalus dolichocephalus 3.908550  2.60131528 -2.59530770
#>  94:      hendersoni      hendersoni 3.859835  1.28963045 -2.06309845
#>  95:     darlingtoni     darlingtoni 4.302036  4.06703170 -1.49378979
#>  96:        aliniger        aliniger 4.036557  0.12323898 -1.67467041
#>  97:      singularis      singularis 4.057997  0.36326789 -0.95422871
#>  98:    chlorocyanus    chlorocyanus 4.275448  0.43906343  1.35403450
#>  99:     coelestinus     coelestinus 4.297965 -0.02721683  0.36875369
#> 100:        occultus        occultus 3.663049  7.92078444 -0.19013968
#>            tip.label       tip.label      SVL   PCI_limbs   PCII_head
#>      PCIII_padwidth_vs_tail PCIV_lamella_num  awesomeness    hostility
#>   1:             -1.0422187      -2.41474228 -0.241651698 -0.173476906
#>   2:             -1.0253056      -2.46331106  0.624468879 -0.500096224
#>   3:             -1.1929572      -2.08743282 -0.427757376  0.480044450
#>   4:              0.1616442      -2.11260585  0.169425977 -0.410812337
#>   5:              0.1693635      -1.37576941 -0.630433775  0.719312962
#>   6:              0.6399320      -1.62629939 -1.754300558  1.412718374
#>   7:              0.3553405      -2.10505922 -0.257638879  0.462708058
#>   8:             -1.3152005      -2.23775137  0.354418772  0.053661422
#>   9:              0.4640379      -1.19666399 -0.228920131  0.820920959
#>  10:              0.6600716      -1.65749549  2.145582200 -0.993730897
#>  11:              0.8679694      -1.56665822 -0.085375433  0.092602997
#>  12:              0.1551290      -1.39796407  0.496114209 -0.544543392
#>  13:              0.6071245      -1.48220834 -0.139145550 -0.310434629
#>  14:              0.2594994       0.10511495  0.167791307  0.676756002
#>  15:              1.5569939       0.93665013  1.483021624 -0.908266532
#>  16:              1.6852579       1.01441926  0.410642801 -0.117462571
#>  17:              1.8543308       0.12882327 -0.088370075  0.465286788
#>  18:             -1.7564495      -1.43952303  0.800931449  0.170290172
#>  19:             -2.3534242      -0.79929053  0.260965443 -0.271697917
#>  20:              1.8824633       1.92392086  1.871367582 -2.029729624
#>  21:              2.2166952       0.94622231 -0.314853533  0.515229477
#>  22:             -1.2588892       0.24029913  0.972494741 -0.371866585
#>  23:             -1.5906715       0.68180888  0.978157532 -0.810754596
#>  24:             -2.2852265      -2.48420519 -0.426336538  0.613014449
#>  25:             -1.6249761      -0.94924727 -0.289982852  0.510241609
#>  26:             -0.1092647      -0.50826932  1.277956446 -1.082760383
#>  27:              1.2175309      -1.21819703  0.830658098 -0.007447027
#>  28:              2.4070756      -0.82666991 -1.538338585  1.633533660
#>  29:              1.7530966      -1.44378878 -0.601304152  0.216500950
#>  30:              1.6898101      -1.57063776 -2.149486418  2.413050748
#>  31:              2.9024438      -1.53259879  0.025584758  0.124044450
#>  32:              2.4922170      -1.03972507  1.973200862 -1.608003643
#>  33:             -2.6138245      -4.30147185  1.788945648 -0.913725183
#>  34:             -2.6721669       1.23487950 -1.253724344  1.308419442
#>  35:             -2.8957227       0.92280509 -0.612226281  0.807322377
#>  36:             -2.4337310       0.62906887 -1.391771586  1.210221860
#>  37:             -2.0659201      -0.74890866  0.848149691 -0.960258609
#>  38:             -1.7589251      -1.71914111  1.456935239 -1.448418148
#>  39:              0.9139707      -0.42075245  0.066715596 -0.290651288
#>  40:             -1.9200701      -2.41993928 -1.457775289  1.096071557
#>  41:             -1.7798344      -2.20379012  0.721747833 -0.969190196
#>  42:             -1.8978249      -2.16718706 -1.944082859  2.147925602
#>  43:             -0.8883209      -2.02129909 -0.158605136 -0.369246555
#>  44:             -0.8299057      -2.09049464  0.905454434 -0.262630999
#>  45:             -1.3884348      -2.19082487  0.257277082  0.029600473
#>  46:             -1.6182845      -2.67106616  2.537213358 -1.420542767
#>  47:             -0.9366696      -1.57843390  0.340077411 -0.690011247
#>  48:             -0.8020303      -1.04842823  1.182343323 -0.167193617
#>  49:             -0.5409516      -0.16220546  2.727286255 -2.487661729
#>  50:             -0.6158464      -0.19473780  1.831370900 -1.331616706
#>  51:             -0.8626956       0.03516802 -0.456651495  0.663302028
#>  52:              1.0228151       1.66539278  2.155800992 -2.494661307
#>  53:             -0.1725921       0.64693336 -0.266699615  0.220410336
#>  54:             -1.0167340      -0.02153073 -0.115917120  0.281153377
#>  55:              1.3476390      -1.64780456  2.156287540 -2.094147109
#>  56:              0.7500647      -1.19909229  0.158971414 -0.350187677
#>  57:              1.3297304      -1.51973745 -1.193993760  1.858148718
#>  58:             -0.1820287      -0.10748930 -0.137962568  0.119982369
#>  59:              2.3331196       0.68195498  1.164592168 -0.865462260
#>  60:              2.2875989      -0.08659903  0.573048475 -0.985921965
#>  61:              2.4409248      -0.07418779 -0.194644197  0.001671326
#>  62:              0.5203703       2.05565172 -0.132309870  0.464556617
#>  63:              0.9781194       2.88538479 -0.074646660  0.148259194
#>  64:              2.3710815       0.33264142  0.007554828 -0.259421287
#>  65:              1.7586477      -1.13869021 -0.620236225  0.330225203
#>  66:              2.1572563      -1.12208954  0.459254083 -0.162682657
#>  67:              2.7442074      -0.52657687  2.153959487 -1.198967281
#>  68:              1.7597408       1.55102509 -0.015197500 -0.020516523
#>  69:              2.7042981      -0.31778878  0.078986096  0.504242392
#>  70:              1.6981742      -0.25962432 -0.431283305  0.321955192
#>  71:              2.2084510       0.94969689 -0.259032167  0.127344278
#>  72:              1.1594671       0.38835687  1.713217290 -1.884387903
#>  73:              0.9491115       1.80146660 -0.299550789  0.846263466
#>  74:             -1.6539883       1.41160228  0.325807029 -0.523648505
#>  75:             -2.8057449       1.68036114  0.244666550 -0.307162858
#>  76:             -1.2178042      -0.55865998  0.340444764 -0.599034616
#>  77:             -2.4795271      -0.03297035  1.327984163 -0.994951781
#>  78:             -2.8189997       0.87176931  0.413447859  0.073949884
#>  79:             -2.4463337       0.70036608 -0.462062012  0.178370302
#>  80:             -3.7238293       1.21810218  0.868463721 -0.287882420
#>  81:             -2.4921192       0.18942396 -1.287008036  1.023037680
#>  82:              0.9791727       2.94194227  0.490363636 -0.505416684
#>  83:              2.1899604       1.75435972  0.126135253  0.545903248
#>  84:              1.1234288       1.74943703 -1.189744988  1.025075387
#>  85:             -1.8467441       0.71173422  0.159291276 -0.024230945
#>  86:              0.1805891       2.66309134  1.821981086 -1.522244645
#>  87:              0.8905492       2.09456880 -0.974406590  1.602507741
#>  88:              0.4296432       2.44067497 -0.018317012 -0.366531140
#>  89:              1.0058546       1.98043577  0.194610912  0.713343825
#>  90:              0.7108046       1.66043194  1.080566180 -1.266611407
#>  91:             -2.2390082       0.13156827  0.628989307  0.146877669
#>  92:             -3.0896949       1.56271877  2.553879072 -2.337986521
#>  93:             -4.1439327       3.67023402  0.739483022 -0.989799476
#>  94:             -3.4656535       2.58336718  0.385454354 -0.713316425
#>  95:             -0.2002259       0.31138087  1.718720076 -1.690079016
#>  96:              2.3871336       0.58486431  0.006869282  0.144815367
#>  97:              1.6562154       1.00756926  1.592083294 -1.701052660
#>  98:              1.9531470       1.80512493 -1.065610146  0.920097708
#>  99:              1.6364316       1.02571762  0.290926638 -0.620965953
#> 100:              2.4922819      -0.09577977 -1.191686980  1.215301402
#>      PCIII_padwidth_vs_tail PCIV_lamella_num  awesomeness    hostility
#>         attitude ecomorph      island
#>   1:  0.64437708       TG        Cuba
#>   2:  0.71289104       TG        Cuba
#>   3: -0.96742634       TG        Cuba
#>   4:  0.19635800       TG        Cuba
#>   5: -1.22282764       TG        Cuba
#>   6:  0.18323450       TG        Cuba
#>   7: -0.27127944       TG        Cuba
#>   8: -0.09389530       GB        Cuba
#>   9: -0.73929564       TG        Cuba
#>  10:  1.05775273       TG        Cuba
#>  11: -0.08130904       TG        Cuba
#>  12:  1.36010631       TG        Cuba
#>  13: -0.50480610       TG        Cuba
#>  14: -0.69460080       CG Puerto Rico
#>  15:  0.72613483       TC     Jamaica
#>  16:  0.70229589       TC Puerto Rico
#>  17: -0.56754896       TW     Jamaica
#>  18:  0.33555714       TG     Jamaica
#>  19:  0.01367143        U     Jamaica
#>  20:  1.02877511       TC Puerto Rico
#>  21: -0.40828716       TC Puerto Rico
#>  22:  1.46109470       GB Puerto Rico
#>  23:  1.39277529       GB Puerto Rico
#>  24: -0.69318262       TG Puerto Rico
#>  25: -0.18543914       GB Puerto Rico
#>  26:  1.37721870       TG Puerto Rico
#>  27:  0.39061232       TG Puerto Rico
#>  28: -1.24247895        T  Hispaniola
#>  29:  0.85182003        T  Hispaniola
#>  30: -2.14378229        T  Hispaniola
#>  31:  0.11758642        T  Hispaniola
#>  32:  0.91461500        T  Hispaniola
#>  33:  0.64025272        U  Hispaniola
#>  34: -0.96110895       GB  Hispaniola
#>  35: -0.82943161       GB  Hispaniola
#>  36: -0.29186894       GB  Hispaniola
#>  37:  0.18291181        U  Hispaniola
#>  38:  0.07531334        U  Hispaniola
#>  39:  0.07286240       TW  Hispaniola
#>  40: -0.87393333       TG  Hispaniola
#>  41:  0.74101761       TG  Hispaniola
#>  42: -2.07465046       TG  Hispaniola
#>  43: -0.26866597       TG  Hispaniola
#>  44:  0.68467864       TG  Hispaniola
#>  45:  0.39422885       TG  Hispaniola
#>  46:  1.71562944       TG  Hispaniola
#>  47:  2.11135720       TG  Hispaniola
#>  48:  0.58279004       TG  Hispaniola
#>  49:  2.78477616       CG  Hispaniola
#>  50:  1.70626854       CG  Hispaniola
#>  51: -0.64625837       CG  Hispaniola
#>  52:  1.11438687        U  Hispaniola
#>  53:  1.46547483        U  Hispaniola
#>  54: -0.97876663       CG Puerto Rico
#>  55:  1.97196598        U        Cuba
#>  56:  0.45987065        U        Cuba
#>  57: -1.43308743        U        Cuba
#>  58:  0.04203987        U        Cuba
#>  59:  0.63326685       TC        Cuba
#>  60:  0.88025286       TC        Cuba
#>  61:  0.16264659       TC        Cuba
#>  62: -0.54956842       TC        Cuba
#>  63: -0.57178426       TC        Cuba
#>  64: -0.62559912        U        Cuba
#>  65: -0.71777358        U        Cuba
#>  66: -0.47694819        U        Cuba
#>  67:  1.51422456        T        Cuba
#>  68:  0.08678021       TW        Cuba
#>  69: -0.40436325       TW  Hispaniola
#>  70: -0.80553288       TW  Hispaniola
#>  71:  0.29597325       TW        Cuba
#>  72:  0.92471752       TW        Cuba
#>  73: -0.96311064       TW        Cuba
#>  74:  0.50405157       GB        Cuba
#>  75:  0.06937286       GB        Cuba
#>  76:  1.78969146       GB        Cuba
#>  77:  0.04668455       GB        Cuba
#>  78: -0.50654925       GB        Cuba
#>  79: -0.01958915       GB        Cuba
#>  80:  0.69946303       GB        Cuba
#>  81: -0.67591746       GB        Cuba
#>  82: -0.20254651        U        Cuba
#>  83: -0.58505782        U        Cuba
#>  84:  0.25994625        U        Cuba
#>  85: -0.31575501        U        Cuba
#>  86:  2.17591511       CG        Cuba
#>  87: -0.42228829       CG        Cuba
#>  88:  1.59163441       CG        Cuba
#>  89: -0.56222110       CG        Cuba
#>  90:  0.66730260       CG        Cuba
#>  91: -0.44097752        U  Hispaniola
#>  92:  2.47075117       GB  Hispaniola
#>  93:  1.09191365       GB  Hispaniola
#>  94: -0.39597920       GB  Hispaniola
#>  95:  1.02145675       TW  Hispaniola
#>  96:  0.29376432       TC  Hispaniola
#>  97:  1.45313446       TC  Hispaniola
#>  98: -0.93721735       TC  Hispaniola
#>  99:  1.28033350       TC  Hispaniola
#> 100:  0.03244860       TW Puerto Rico
#>         attitude ecomorph      island